refactor: move enums to parent folder
