+++ src/common/enum/IndustriesEnum.ts
